+++ demo_site/urls.py
@@ -3,7 +3,7 @@ from django.urls import path
 from django.views.generic import RedirectView
 
 urlpatterns = [
-    path('', RedirectView.as_view(url='qr-code-demo/', permanent=True)),
-    path('qr-code-demo/', include('qr_code_demo.urls', namespace='qr_code_demo')),
-    path('qr-code/', include('qr_code.urls', namespace='qr_code')),
+    path("", RedirectView.as_view(url="qr-code-demo/", permanent=True)),
+    path("qr-code-demo/", include("qr_code_demo.urls", namespace="qr_code_demo")),
+    path("qr-code/", include("qr_code.urls", namespace="qr_code")),
 ]