+++ qr_code/qrcode/__init__.py
@@ -1,0 +1,1 @@
+PYDANTIC_CONFIG = dict(arbitrary_types_allowed=True)